a2a_multiagent/pyproject.toml
[project]
name = "a2a-multiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncclick>=8.1.8",
    "click>=8.2.1",
    "fastapi>=0.115.12",
    "google-adk>=1.2.1",
    "google-genai>=1.19.0",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "starlette>=0.46.2",
    "tzlocal>=5.3.1",
    "uvicorn>=0.34.3",
]
